requirements.txt
typer
pydantic
pyyaml
rich
requests
pytest
python-dotenv
